wolt_checker/requirements.txt
requests~=2.27
python-telegram-bot~=13.11
firebase_admin~=5.2
